+++ src/commands/services/mod.rs
@@ -1,3 +1,4 @@
+pub mod join;
 use serenity::{model::channel::Message, Result as SerenityResult};
 
 /// Checks that a message successfully sent; if not, then logs why to stdout.
