+++ src/model/mod.rs
@@ -1,0 +1,2 @@
+mod voice;
+pub use voice::Voice;
